+++ cucumber-expressions/ruby/spec/cucumber/cucumber_expressions/tree_regexp_spec.rb
@@ -84,6 +84,13 @@ module Cucumber
         group = tr.match("hello")
         expect(group.value).to eq("hello")
       end
+
+      it 'throws an error when there are named capture groups because they are buggy in Ruby' do
+        # https://github.com/cucumber/cucumber/issues/329
+        expect {
+          TreeRegexp.new(/^I am a person( named "(?<first_name>.+) (?<last_name>.+)")?$/)
+        }.to raise_error(/Named capture groups are not supported/)
+      end
     end
   end
 end